Validacion de structura email y correccion en user_module

+++ src/user/user.service.ts
@@ -1,10 +1,14 @@
-import { Injectable } from '@nestjs/common';
+import { ForbiddenException, Injectable } from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { EditUserDto } from './dto';
 
 @Injectable()
 export class UserService {
   constructor(private prisma: PrismaService) {}
+
+  async findAll() {
+    return await this.prisma.usuario.findMany();
+  }
 
   async update(id: number, informacion: EditUserDto): Promise<EditUserDto> {
     return await this.prisma.usuario.update({
@@ -25,10 +29,20 @@ export class UserService {
     });
   }
 
-  remove(id: number) {
-    return this.prisma.usuario.delete({
+  async remove(id: number) {
+    const user = await this.prisma.usuario.findUnique({
       where: {
         id: typeof id === 'number' ? id : Number.parseInt(id),
+      },
+    });
+
+    // if user does not exist throw exception
+    if (!user) throw new ForbiddenException('User Not Found on DB');
+
+    //send back the user
+    return this.prisma.usuario.delete({
+      where: {
+        id: user.id,
       },
     });
   }
